Move some functions to commands/lock and commands/pip
