+++ src/lib.rs
@@ -18,7 +18,6 @@ mod win_job;
 use crate::cli::commands;
 pub use crate::cli::syntax::Command;
 use crate::cli::syntax::SubCommand;
-use crate::dependencies::FrozenDependency;
 pub use crate::error::*;
 use crate::lock::BumpType;
 use crate::operations::UpdateOptions;
@@ -112,7 +111,7 @@ pub fn run_cmd(cmd: Command) -> Result<(), Error> {
                 python_version: python_version.clone(),
                 sys_platform: sys_platform.clone(),
             };
-            update_lock(&context?, update_options)
+            commands::update_lock(&context?, update_options)
         }
         SubCommand::BumpInLock { name, version, git } => {
             let bump_type = if *git {
@@ -120,7 +119,7 @@ pub fn run_cmd(cmd: Command) -> Result<(), Error> {
             } else {
                 BumpType::Simple
             };
-            bump_in_lock(&context?, name, version, bump_type)
+            commands::bump_in_lock(&context?, name, version, bump_type)
         }
         SubCommand::Run { ref cmd, no_exec } => {
             if *no_exec {
@@ -136,52 +135,12 @@ pub fn run_cmd(cmd: Command) -> Result<(), Error> {
 
         SubCommand::Tidy {} => tidy(&cmd, &context?),
 
-        SubCommand::UpgradePip {} => upgrade_pip(&context?),
+        SubCommand::UpgradePip {} => commands::upgrade_pip(&context?),
     }
 }
 
 fn process_scripts(context: &Context, mode: ProcessScriptsMode) -> Result<(), Error> {
     operations::scripts::process(&context.paths, mode)
-}
-
-/// (Re)generate the lock file
-//
-// Notes:
-//
-// * Abort if `setup.py` is not found
-// * Create the virtualenv if required
-// * Always upgrade pip :
-//    * If that fails, we know if the virtualenv is broken
-//    * Also, we know sure that `pip` can handle all the options
-//      (such as `--local`, `--exclude-editable`) we use in the other functions
-// * The path of the lock file is computed by PathsResolver.
-//     See PathsResolver.paths() for details
-fn update_lock(context: &Context, update_options: operations::UpdateOptions) -> Result<(), Error> {
-    print_info_1("Updating lock");
-    let Context { paths, .. } = context;
-    if !&paths.setup_py.exists() {
-        return Err(Error::MissingSetupPy {});
-    }
-    commands::ensure_venv(&context)?;
-    upgrade_pip(&context)?;
-    install_editable(&context)?;
-    let metadata = metadata(&context);
-    let frozen_deps = get_frozen_deps(&context)?;
-    let lock_path = &paths.lock;
-    operations::lock::update(lock_path, frozen_deps, update_options, &metadata)
-}
-
-/// Bump a dependency in the lock file
-fn bump_in_lock(
-    context: &Context,
-    name: &str,
-    version: &str,
-    bump_type: BumpType,
-) -> Result<(), Error> {
-    print_info_1(&format!("Bumping {} to {} ...", name, version));
-    let metadata = metadata(&context);
-    let Context { paths, .. } = context;
-    operations::lock::bump(&paths.lock, name, version, bump_type, &metadata)
 }
 
 // Re-generate a clean lock:
@@ -198,9 +157,9 @@ fn tidy(cmd: &Command, context: &Context) -> Result<(), Error> {
     // deleted the python we used to clean the previous virtualenv
     let context = get_context(&cmd)?;
     commands::create_venv(&context)?;
-    install_editable_with_constraint(&context)?;
-    let metadata = metadata(&context);
-    let frozen_deps = get_frozen_deps(&context)?;
+    commands::install_editable_with_constraint(&context)?;
+    let metadata = commands::metadata(&context);
+    let frozen_deps = commands::get_frozen_deps(&context)?;
     let Context { paths, .. } = context;
     operations::lock::tidy(&paths.lock, frozen_deps, &metadata)
 }
@@ -243,68 +202,6 @@ fn show_venv_bin_path(context: &Context) -> Result<(), Error> {
     Ok(())
 }
 
-fn upgrade_pip(context: &Context) -> Result<(), Error> {
-    let Context { venv_runner, .. } = context;
-    print_info_2("Upgrading pip");
-    let cmd = &["python", "-m", "pip", "install", "pip", "--upgrade"];
-    venv_runner.run(cmd).map_err(|_| Error::UpgradePipError {})
-}
-
-fn install_editable(context: &Context) -> Result<(), Error> {
-    let Context {
-        settings,
-        venv_runner,
-        ..
-    } = context;
-    let mut message = "Installing deps from setup.py".to_string();
-    if settings.production {
-        message.push_str(" using 'prod' extra dependencies");
-    } else {
-        message.push_str(" using 'dev' extra dependencies");
-    }
-    print_info_2(&message);
-    let cmd = get_install_editable_cmd(&context);
-    venv_runner.run(&cmd)
-}
-
-fn install_editable_with_constraint(context: &Context) -> Result<(), Error> {
-    let Context {
-        paths, venv_runner, ..
-    } = context;
-    let lock_path = &paths.lock;
-    let message = format!(
-        "Installing deps from setup.py, constrained by {}",
-        lock_path.display()
-    );
-    print_info_2(&message);
-    let lock_path_str = lock_path.to_string_lossy();
-    let mut cmd = get_install_editable_cmd(&context).to_vec();
-    cmd.extend(&["--constraint", &lock_path_str]);
-    venv_runner.run(&cmd)
-}
-
-fn get_install_editable_cmd(context: &Context) -> [&str; 6] {
-    let Context { settings, .. } = context;
-    let extra = if settings.production {
-        ".[prod]"
-    } else {
-        ".[dev]"
-    };
-    ["python", "-m", "pip", "install", "--editable", extra]
-}
-
-fn metadata(context: &Context) -> Metadata {
-    let Context { python_info, .. } = context;
-    let dmenv_version = env!("CARGO_PKG_VERSION");
-    let python_platform = &python_info.platform;
-    let python_version = &python_info.version;
-    Metadata {
-        dmenv_version: dmenv_version.to_string(),
-        python_platform: python_platform.to_string(),
-        python_version: python_version.to_string(),
-    }
-}
-
 fn look_up_for_project_path() -> Result<PathBuf, Error> {
     let mut candidate = std::env::current_dir()
         .map_err(|e| new_error(format!("Could not get current directory: {}", e)))?;
@@ -324,36 +221,6 @@ fn look_up_for_project_path() -> Result<PathBuf, Error> {
             }
         }
     }
-}
-
-/// Get the list of the *actual* deps in the virtualenv by calling `pip freeze`.
-fn get_frozen_deps(context: &Context) -> Result<Vec<FrozenDependency>, Error> {
-    let freeze_output = run_pip_freeze(&context)?;
-    // First, collect all the `pip freeze` lines into frozen dependencies
-    let deps: Result<Vec<_>, _> = freeze_output
-        .lines()
-        .map(|x| FrozenDependency::from_string(x.into()))
-        .collect();
-    let deps = deps?;
-    // Then filter out pkg-resources: this works around a Debian bug in pip:
-    // https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=871790
-    let res: Vec<_> = deps
-        .into_iter()
-        .filter(|x| x.name != "pkg-resources")
-        .collect();
-    Ok(res)
-}
-
-fn run_pip_freeze(context: &Context) -> Result<String, Error> {
-    let Context { venv_runner, .. } = context;
-    #[rustfmt::skip]
-        let cmd = &[
-            "python", "-m", "pip", "freeze",
-            "--exclude-editable",
-            "--all",
-            "--local",
-        ];
-    venv_runner.get_output(cmd)
 }
 
 /// Run a program from the virtualenv, making sure it dies
